Refactor orm typing

+++ app/orm/balance.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 from sqlalchemy.orm import Mapped, mapped_column
 
-from .core import ORMModel
+from .core import ORMModel, types
 
 
 class BalanceModel(ORMModel):
-    id: Mapped[int] = mapped_column(unique=True, primary_key=True)
-    user_id: Mapped[int] = mapped_column(unique=True)
-    balance: Mapped[int] = mapped_column(default=15000)
+    id: Mapped[types.BigInt] = mapped_column(unique=True, primary_key=True)
+    user_id: Mapped[types.BigInt] = mapped_column(unique=True)
+    balance: Mapped[types.BigInt] = mapped_column(default=15000)
